Added payment for order
